food recognation/requirements.txt
tensorflow>=2.15.0
streamlit
pandas
numpy
matplotlib
seaborn
pillow
scikit-learn
opencv-python
